refactored course summary enrollment into enrollment.py, addressed PR feedback
