+++ setup.cfg
@@ -70,7 +70,6 @@ edx.analytics.tasks =
     obfuscation = edx.analytics.tasks.obfuscation:ObfuscatedCourseTask
     push_to_vertica_lms_courseware_link_clicked = edx.analytics.tasks.lms_courseware_link_clicked:PushToVerticaLMSCoursewareLinkClickedTask
     load-course-catalog = edx.analytics.tasks.load_internal_reporting_course_catalog:PullCourseCatalogAPIData
-    course-summary-enrollment= edx.analytics.tasks.course_summary_enrollment:CourseSummaryEnrollmentWrapperTask
 
 
 mapreduce.engine =
